// src/features/auth/schemas/AuthSchema.ts
import * as yup from "yup";

export const schema = yup
  .object({
    firstName: yup.string().min(2).required(),
    lastName: yup.string().min(2).required(),
    email: yup.string().required(),
    password: yup.string().min(6).required(),
  })
  .required();

export const loginSchema = yup
  .object({
    email: yup.string().required(),
    password: yup.string().min(6).required(),
  })
  .required();